requirements.txt
streamlit
pillow
google-generativeai


pip install -r requirements.txt